split ldap code
