+++ internal/ldap/entries.go
@@ -1,6 +1,9 @@
 package ldap
 
-import "github.com/ps78674/gorestldap/internal/data"
+import (
+	"github.com/google/uuid"
+	"github.com/ps78674/gorestldap/internal/data"
+)
 
 // GetEntries returns configured entries
 func GetEntries(baseDN, usersOUName, groupsOUName string) *data.Entries {
@@ -40,4 +43,9 @@ func GetEntries(baseDN, usersOUName, groupsOUName string) *data.Entries {
 		Domain: domain,
 		OUs:    ous,
 	}
+}
+
+// newEntryUUID creates uuid5 from NameSpaceOID and entry name
+func newEntryUUID(name string) string {
+	return uuid.NewSHA1(uuid.NameSpaceOID, []byte(name)).String()
 }
